Refactor: Improve project structure and code organization
This commit introduces several changes to enhance the project's structure, maintainability, and adherence to good practices.

Key changes include:

1.  **Component Reorganization**:
    -   Components in `app/components/` have been moved into more specific subdirectories: `features/`, `ui/`, and `utils/`.
    -   Import paths have been updated accordingly.

2.  **Asset Directory Restructuring**:
    -   The `public/assets/` directory has been reorganized. CSS, documents (files), and images are now in `public/css/`, `public/documents/`, and `public/images/` respectively.
    -   All references to these assets in `app/layout.tsx` and `app/data/content.json` have been updated.

3.  **Styling Readability**:
    -   Added comments to the global stylesheet (`public/css/style.css`) to delineate sections of styles that are primarily associated with specific components, improving navigability.

4.  **TypeScript Enhancements**:
    -   Centralized the `Project` interface into a new `app/types.ts` file.
    -   Updated `ProjectList.tsx` to import the `Project` interface from the shared location.
    -   Reviewed overall TypeScript usage for basic correctness.

5.  **Build Verification**:
    -   The application was successfully built after these changes, ensuring no build-time errors were introduced.

These changes aim to make the codebase easier to understand, navigate, and scale in the future.
